services/tgbot/requirements.txt
Flask==3.0.2
PyYAML==6.0.1
cryptography==42.0.5
python-telegram-bot==21.0.1